+++ diesel/src/pg/query_builder/query_fragment_impls.rs
@@ -9,20 +9,3 @@ impl QueryFragment<Pg> for ForUpdateClause {
         Ok(())
     }
 }
-
-#[derive(Debug, Clone, Copy)]
-pub struct DistinctOnClause<T>(pub(crate) T);
-
-impl<T> QueryFragment<Pg> for DistinctOnClause<T>
-where
-    T: QueryFragment<Pg>,
-{
-    fn walk_ast(&self, mut out: AstPass<Pg>) -> QueryResult<()> {
-        out.push_sql("DISTINCT ON(");
-        self.0.walk_ast(out.reborrow())?;
-        out.push_sql(")");
-        Ok(())
-    }
-}
-
-impl_query_id!(DistinctOnClause<T>);
